+++ src/index.ts
@@ -186,6 +186,7 @@ class TheAlgorithm {
             const userData = allResponses.shift();
             this.mutedAccounts = userData.mutedAccounts;
             this.serverSideFilters = userData.serverSideFilters;
+            this.mastodonServers = await MastoApi.instance.getMastodonServersInfo();  // Should load from storage
             // Pull followed accounts and tags from the scorers
             this.followedAccounts = Account.buildAccountNames(this.mentionsFollowedScorer.followedAccounts);
             this.followedTags = this.followedTagsScorer.requiredData;
@@ -193,7 +194,7 @@ class TheAlgorithm {
 
         this.logTootCounts(newToots, homeToots);
         // Remove stuff already retooted, invalid future timestamps, nulls, etc.
-        let cleanNewToots = newToots.filter(toot => toot.isValidForFeed(this));
+        const cleanNewToots = newToots.filter(toot => toot.isValidForFeed(this));
         const numRemoved = newToots.length - cleanNewToots.length;
         console.log(`Removed ${numRemoved} invalid toots leaving ${cleanNewToots.length}`);
 
@@ -201,12 +202,6 @@ class TheAlgorithm {
         this.feed = cleanFeed.slice(0, Storage.getConfig().maxNumCachedToots);
         this.initializeFiltersWithSummaryInfo();
         this.maybeGetMoreToots(homeToots, numTimelineToots);  // Called asynchronously
-
-        // TODO: this is a hacky way to preserve the server domains...
-        if (Object.keys(this.mastodonServers).length == 0) {
-            this.mastodonServers = await MastoApi.instance.getMastodonServersInfo();
-        }
-
         return this.scoreFeed.bind(this)();
     }